ONB-211-2 chore: adpater-in:support 모듈 분리
