+++ adapter-in/support/auth/src/main/kotlin/com/yapp/bol/auth/dto/LoginRequest.kt
@@ -1,0 +1,8 @@
+package com.yapp.bol.auth.dto
+
+import com.yapp.bol.auth.LoginType
+
+data class LoginRequest(
+    val type: LoginType,
+    val token: String,
+)
