+++ adapter-in/support/oas/src/test/kotlin/com/yapp/bol/base/DocumentFieldType.kt
@@ -1,0 +1,20 @@
+package com.yapp.bol.base
+
+import org.springframework.restdocs.payload.JsonFieldType
+import kotlin.reflect.KClass
+
+sealed class DocumentFieldType(
+    val type: JsonFieldType,
+)
+
+object ARRAY : DocumentFieldType(JsonFieldType.ARRAY)
+object BOOLEAN : DocumentFieldType(JsonFieldType.BOOLEAN)
+object OBJECT : DocumentFieldType(JsonFieldType.OBJECT)
+object NUMBER : DocumentFieldType(JsonFieldType.NUMBER)
+object NULL : DocumentFieldType(JsonFieldType.NULL)
+object STRING : DocumentFieldType(JsonFieldType.STRING)
+object ANY : DocumentFieldType(JsonFieldType.VARIES)
+
+data class ENUM<T : Enum<T>>(val enums: List<T>) : DocumentFieldType(JsonFieldType.STRING) {
+    constructor(clazz: KClass<T>) : this(clazz.java.enumConstants.asList()) // (1)
+}
