+++ adapter-in/support/oas/src/test/kotlin/com/yapp/bol/base/OpenApiTag.kt
@@ -1,0 +1,16 @@
+package com.yapp.bol.base
+
+enum class OpenApiTag(
+    val value: String,
+) {
+    TEST("test"),
+    AUTH("Auth"),
+    TERMS("Terms"),
+    SETTING("Setting"),
+    FILE("File"),
+    GAME("Game"),
+    GROUP("Group"),
+    MEMBER("Member"),
+    USER("User"),
+    MATCH("Match"),
+}
