+++ adapter-in/web/build.gradle.kts
@@ -13,6 +13,8 @@ plugins {
 dependencies {
     val springVersion by properties
     implementation(project(":adapter-in:admin"))
+    implementation(project(":adapter-in:support:auth"))
+    implementation(project(":adapter-in:support:oas"))
 
     implementation(project(":support:logging"))
     implementation(project(":support:yaml"))
@@ -20,14 +22,11 @@ dependencies {
 
     implementation("org.springframework.boot:spring-boot-starter-web:$springVersion")
     implementation("org.springframework.boot:spring-boot-starter-aop:$springVersion")
-    implementation("org.springframework.boot:spring-boot-starter-security:$springVersion")
 
+    testImplementation(project(":adapter-in:support:auth", "testArchive"))
+    testImplementation(project(":adapter-in:support:oas", "testArchive"))
     testImplementation(project(":domain", "testArchive"))
     testImplementation("org.springframework.boot:spring-boot-starter-test:$springVersion")
-    testImplementation("org.springframework.restdocs:spring-restdocs-mockmvc:3.0.0")
-    testImplementation("com.epages:restdocs-api-spec-mockmvc:0.18.0")
-    swaggerCodegen("io.swagger.codegen.v3:swagger-codegen-cli:3.0.42")
-    swaggerUI("org.webjars:swagger-ui:4.18.2")
 }
 
 tasks {
