--- adapter-in/web/src/main/kotlin/com/yapp/bol/auth/dto/LoginRequest.kt
@@ -1,8 +1,0 @@
-package com.yapp.bol.auth.dto
-
-import com.yapp.bol.auth.LoginType
-
-data class LoginRequest(
-    val type: LoginType,
-    val token: String,
-)
