--- adapter-in/web/src/main/kotlin/com/yapp/bol/auth/dto/LoginResponse.kt
@@ -1,9 +1,0 @@
-package com.yapp.bol.auth.dto
-
-import com.fasterxml.jackson.annotation.JsonInclude
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-data class LoginResponse(
-    val accessToken: String,
-    val refreshToken: String?,
-)
