--- adapter-in/web/src/main/kotlin/com/yapp/bol/auth/security/SecurityExceptionHandler.kt
@@ -1,32 +1,0 @@
-package com.yapp.bol.auth.security
-
-import jakarta.servlet.http.HttpServletRequest
-import jakarta.servlet.http.HttpServletResponse
-import org.springframework.beans.factory.annotation.Qualifier
-import org.springframework.security.access.AccessDeniedException
-import org.springframework.security.core.AuthenticationException
-import org.springframework.security.web.AuthenticationEntryPoint
-import org.springframework.security.web.access.AccessDeniedHandler
-import org.springframework.stereotype.Component
-import org.springframework.web.servlet.HandlerExceptionResolver
-
-@Component
-class SecurityExceptionHandler(
-    @Qualifier("handlerExceptionResolver") private val handler: HandlerExceptionResolver,
-) : AuthenticationEntryPoint, AccessDeniedHandler {
-    override fun commence(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        authException: AuthenticationException,
-    ) {
-        handler.resolveException(request, response, null, authException)
-    }
-
-    override fun handle(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        accessDeniedException: AccessDeniedException,
-    ) {
-        handler.resolveException(request, response, null, accessDeniedException)
-    }
-}
