+++ adapter-in/web/src/test/kotlin/com/yapp/bol/base/WebControllerTest.kt
@@ -1,0 +1,7 @@
+package com.yapp.bol.base
+
+import com.yapp.bol.ExceptionHandler
+
+abstract class WebControllerTest : BaseControllerTest(
+    controllerAdvices = listOf(ExceptionHandler()),
+)
